Refactor JWT token utility and add middleware for token verification
