+++ backend/src/api/userRoutes.ts
@@ -3,7 +3,7 @@ import { registerUser, retrieveUser } from "../db/userStore";
 import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
 import { JWT_SECRET } from "../config";
-import { newToken } from "../utils/jwtTokens";
+import { makeJwt } from "../utils/auth";
 
 const routes = express.Router();
 const tokenExpiry = 30 * 60 * 1000;
@@ -31,7 +31,7 @@ routes.post('/login', async (req: Request, res: Response) => {
         }
         const isMatch = await bcrypt.compare(password, foundUser.password);
         if (isMatch) {
-            const token = await newToken({
+            const token = await makeJwt({
                 _id: foundUser._id.toString(),
                 username: foundUser.username,
                 exp: tokenExpiry
