+++ backend/test/utils/auth.test.ts
@@ -1,7 +1,7 @@
 import 'mocha';
 import { expect } from 'chai';
 import jwt from 'jsonwebtoken';
-import { newToken } from '../../src/utils/jwtTokens';
+import { makeJwt } from '../../src/utils/auth';
 import { JWT_SECRET } from '../../src/config';
 
 const testUser = {
@@ -12,19 +12,19 @@ const testUser = {
 
 describe('JWT Token Utility Tests', () => {
     it('should generate a valid token', async () => {
-        const token = await newToken(testUser, JWT_SECRET);
+        const token = await makeJwt(testUser, JWT_SECRET);
         expect(token).to.not.be.empty;
     });
 
     it('should contain the correct payload', async () => {
-        const token = await newToken(testUser, JWT_SECRET);
+        const token = await makeJwt(testUser, JWT_SECRET);
         const decoded = jwt.verify(token, JWT_SECRET) as any;
         expect(decoded).to.have.property('_id', testUser._id);
         expect(decoded).to.have.property('username', testUser.username);
     });
 
     it('should set the expiration correctly', async () => {
-        const token = await newToken(testUser, JWT_SECRET);
+        const token = await makeJwt(testUser, JWT_SECRET);
         const decoded = jwt.verify(token, JWT_SECRET) as any;
         expect(decoded).to.have.property('exp');
         expect(decoded.exp).to.be.above(0);
@@ -32,7 +32,7 @@ describe('JWT Token Utility Tests', () => {
 
     it('should throw an error with empty secret', async () => {
         try {
-            await newToken(testUser, '');
+            await makeJwt(testUser, '');
             expect.fail('Should have thrown an error');
         } catch (err) {
             expect(err).to.exist;
@@ -44,7 +44,7 @@ describe('JWT Token Utility Tests', () => {
     it('should throw an error with missing payload fields', async () => {
         const incompleteUser = { _id: '12345' };
         try {
-            await newToken(incompleteUser as any, JWT_SECRET);
+            await makeJwt(incompleteUser as any, JWT_SECRET);
             expect.fail('Should have thrown an error');
         } catch (err) {
             expect(err).to.exist;
